emon_Pulse: change daily calculation to be based on startofDay pulse count. Fix unsigned short to unsigned long on Pulse internal count

+++ emon_JeeNodePulse/emon_JeeNodePulse.ino
@@ -19,7 +19,7 @@
 
 RF12Init rf12Init = { PULSE_JEENODE, RF12_915MHZ, FEWINGS_MONITOR_GROUP };
 
-volatile unsigned short	g_pulseCount[NUM_PINS]		= { 0,0,0,0 };	//pulses since recording started
+volatile unsigned long 	g_pulseCount[NUM_PINS]	= { 0,0,0,0 };	//pulses since recording started
 volatile unsigned long	g_lastTick[NUM_PINS]		= { 0,0,0,0 };		//millis() value at last pulse
 volatile unsigned long	g_period[NUM_PINS]			= { 0,0,0,0 };		//ms between last two pulses
 const		double					g_pulsePerWH[NUM_PINS]	= { 2.0,2.0,0.4,1.0};		//number of pulses per wH for each input. Some are 2, some are 1, some are 0.4
